Polyglot V3 updates
Update the interface to work with Polyglot version 3.

- Change how initial configuration is processed. Instead of getting
  JSON object via stdin we now get a base64 encoded JSON string in
  the environment variable PG3INIT.
- Custom Parameters are now stored with a key 'customparams'.

The MQTT topics are a bit different so update those as well.

When sending messages to Polyglot via MQTT, we now need to include
a message type.  The various messages are mapped to a small list of
types so change sendMessage to take the type.