Communicate properly with PG3
PG3 has differences in the way it communicates with node servers
so fix up the interface to account for those differences.

The way PG3 handles connection status is different, it no longer
sends a connected message to the node server (this means the
interface shouldn't look for that, i.e. the polyglotconnected flag).

PG3 doesn't automatically send initial configuration info.  Add
a function to query PG3 for the initial config info after the MQTT
connection has been established.

The format of the messages for things like addnode, customparams,
notices, etc. are different so update the message format to match.